Make images suitable for connection with salt-shell
To be able to connect the image to a separate salt-shell image, which runs sshd
and makes it possible to execute salt commands in a different container, some
more directories had to be moved to the data volume. As well, the salt socket
directory and pid file have to be shared with the salt-shell.